+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.8.2] - Maintenance & Optimisation
 - reduced delete executor to just one mutable borrow
 - added delete_lots test just in case
+- IndexTuple::HAS_INDEXES trait constant so you can easily skip indexes if needed
+- db::Db is now passed to fixtures and not the SaveExecutor
 
 ## [0.8.1] - Internal Changes
 - added CanisterKind, StoreKind, IndexKind so that we can declare the static types